backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic<2.0.0
python-dotenv
psycopg2-binary
asyncpg
alembic
bcrypt==4.0.1
cryptography
email-validator


